EmployeeDropdown complete
Currently I am just displaying one and two as the options. This would be changed as i figure out how to adjust the select options depending on its usage .
It is rendered on the task page for testing.
If you want to adjust the styling try the style attribute of the select or add your styling to the less file i made to style the dropdown